move some functions around
